+++ test/snapshot/core/compound_assign/intended/main.go
@@ -6,12 +6,12 @@ func main() {
 	i := 1
 	_ = i
 	now := func() int {
-		i = (i + 2)
+		i = int(int32((int32(i) + int32(2))))
 		return i
 	}()
 	_ = now
 	hxrt.Println(hxrt.StdString(now))
 	hxrt.Println(hxrt.StdString(i))
-	i = (i * 4)
+	i = int(int32((int32(i) * int32(4))))
 	hxrt.Println(hxrt.StdString(i))
 }
