+++ test/snapshot/core/function_values/intended/main.go
@@ -4,11 +4,11 @@ import "snapshot/hxrt"
 
 func main() {
 	add := func(a int, b int) int {
-		return (a + b)
+		return int(int32((int32(a) + int32(b))))
 	}
 	_ = add
 	mul := func(v int) int {
-		return (v * 3)
+		return int(int32((int32(v) * int32(3))))
 	}
 	_ = mul
 	hxrt.Println(twice(5))
@@ -17,5 +17,5 @@ func main() {
 }
 
 func twice(value int) int {
-	return (value * 2)
+	return int(int32((int32(value) * int32(2))))
 }
