+++ test/snapshot/core/locals_assign/intended/main.go
@@ -5,7 +5,7 @@ import "snapshot/hxrt"
 func main() {
 	total := 1
 	_ = total
-	total = (total + 4)
+	total = int(int32((int32(total) + int32(4))))
 	flag := false
 	_ = flag
 	flag = (total > 3)
