+++ test/snapshot/core/loop_break_continue/intended/main.go
@@ -8,14 +8,14 @@ func main() {
 	sum := 0
 	_ = sum
 	for i < 10 {
-		i = (i + 1)
-		if (i % 2) == 0 {
+		i = int(int32((i + 1)))
+		if int(int32((int32(i) % int32(2)))) == 0 {
 			continue
 		}
 		if i > 7 {
 			break
 		}
-		sum = (sum + i)
+		sum = int(int32((int32(sum) + int32(i))))
 	}
 	hxrt.Println(hxrt.StdString(sum))
 	hxrt.Println(hxrt.StdString(i))
