+++ test/snapshot/core/varargs_foreach/intended/main.go
@@ -20,12 +20,12 @@ func sum(values []int) int {
 		this1 := _g_args
 		_ = this1
 		hx_post_1 := _g_current
-		_g_current = (_g_current + 1)
+		_g_current = int(int32((_g_current + 1)))
 		index := hx_post_1
 		_ = index
 		value := this1[index]
 		_ = value
-		total = (total + value)
+		total = int(int32((int32(total) + int32(value))))
 	}
 	return total
 }
